\name{ma.coxph.h}
\alias{ma.coxph.h}
\title{helper function}
\description{Rogues, swindlers, rascals: Film history has probably created a variety of Krimininalikonen, such as Bonnie and Clyde, Bender Bending Rodriguez, Frank William Abagnale Junior, Doctor Kimble, Jesse James, Bill the Butcher and others. As strong and complex as these characters may be, as much as their actors give expression to them, and as much as they are known around the world, they are so insignificant compared to the only real rogue: Tierpark-Toni! And if the Monaco had not been there, then...}
\keyword{internal}