\name{ms.h}
\alias{ms.h}
\title{helper function}
\description{A last feeling of tender sadness, faunophilic memories of green steppes rise to melancholy, a pastureland of the decadent word archivist and acrobat escapes the sentiment; oh melancholy - where are you leading us now? Thus, with a drop of wormwood, we finally breathe the farewell of the true being and revel in the statues of self-satisfied words; and yet there is hope: even though our sadness laments the futility of the present, our thoughts are inexorably controlling the sun: yes, a source of inspiration will surely pour over us in the future ...}
\keyword{internal}