\name{mami-package}
\alias{mami-package}
\alias{MAMI}
\docType{package}
\title{
Model averaging (and model selection) after multiple imputation
}
\description{
Performs model selection/averaging on multiply imputed data sets and combines the resulting estimates. The package also provides access to less frequently used 
model averaging techniques and offers integrated bootstrap estimation. The main function is \code{"mami"}.
}

\details{
\tabular{ll}{
Package: \tab mami\cr
Type: \tab Package\cr
Version: \tab 0.9.5\cr
Date: \tab 2016-10-16\cr
License: \tab GPL-2\cr
Imports: \tab R (>= 3.3.2), lasso2, quadprog, corpcor, boot, MASS, survival, MuMIn, zoo, lme4, Amelia, BMA\cr
}

}
\author{
Michael Schomaker

Maintainer: Michael Schomaker <michael.schomaker@uct.ac.za>
}
\references{
Schomaker, M., Heumann, C. (2014) \emph{Model Selection and Model Averaging after Multiple Imputation},
Computational Statistics & Data Analysis, 71:758-770
}
\keyword{Model Averaging}

